Restructure tabs. Split each into files
