--- client/src/plus/common/plus-tab.tsx
@@ -1,325 +1,0 @@
-import { useEffect, useRef, useState } from "react";
-import { useUIStatus } from "../../ui-context";
-import { Button } from "../../ui/atoms/button";
-import { CollectionListItem } from "../collections/collection-list-item";
-import WatchedCardListItem from "../icon-card";
-import SearchFilter from "../search-filter";
-import {
-  deleteItemById,
-  deleteItemsById,
-  starItem,
-  starItemsById,
-  undoDeleteItemById,
-  unstarItemsById,
-  unwatchItemsByUrls,
-  updateCollectionItem,
-  updateDeleteCollectionItem,
-} from "../notifications/api";
-import NotificationCardListItem from "../notifications/notification-card-list-item";
-import SelectedNotificationsBar from "../notifications/notification-select";
-import { FILTERS, SORTS, TabVariant, TAB_INFO } from "../notifications/tabs";
-import { useVisibilityChangeListener } from "../notifications/utils";
-import { BookmarkData } from "../collections";
-
-export function PlusTab({
-  currentTab,
-  selectedTerms,
-  selectedFilter,
-  selectedSort,
-  showSelectToolbar = true,
-  data,
-  offset,
-  setOffset,
-  hasMore,
-  setFrequentlyUsed,
-}) {
-  const { setToastData } = useUIStatus();
-  const listRef = useRef<Array<any>>([]);
-  const [list, setList] = useState<Array<any>>([]);
-  const [selectAllChecked, setSelectAllChecked] = useState(false);
-  const [editOptions, setEditOptions] = useState({
-    starEnabled: false,
-    unstarEnabled: false,
-    deleteEnabled: false,
-    unwatchEnabled: false,
-  });
-  listRef.current = list;
-  // Uncheck and clear list on tab or filter change
-  useEffect(() => {
-    setSelectAllChecked(false);
-    setList([]);
-  }, [currentTab, selectedFilter, selectedSort, selectedTerms]);
-
-  useVisibilityChangeListener();
-
-  useEffect(() => {
-    if (data && !!data.items) {
-      setList([
-        ...listRef.current,
-        ...data.items.map((item) => {
-          return { ...item, checked: false };
-        }),
-      ]);
-    }
-  }, [data]);
-
-  const deleteItem = async (item) => {
-    await deleteItemById(data.csrfmiddlewaretoken, item.id);
-    const listWithDelete = list.filter((v) => v.id !== item.id);
-    setList(listWithDelete);
-    setToastData({
-      mainText: `${item.title} removed from your collection`,
-      shortText: "Article removed",
-      buttonText: "Undo",
-      buttonHandler: async () => {
-        await undoDeleteItemById((data as any).csrfmiddlewaretoken, item.id);
-        setToastData(null);
-      },
-    });
-  };
-
-  const deleteMany = async () => {
-    const toDelete = list.filter((v) => v.checked).map((i) => i.id);
-    await deleteItemsById(data.csrfmiddlewaretoken, toDelete);
-    const updated = list.filter((v) => !v.checked);
-    setList(updated);
-  };
-
-  const toggleStarItem = async (item) => {
-    await starItem(data.csrfmiddlewaretoken, item.id);
-    //Local updates
-    if (currentTab === TabVariant.STARRED) {
-      setList([...list.filter((v) => v.id !== item.id)]);
-    } else {
-      const updated = list.map((v) => {
-        if (v.id === item.id) {
-          v.starred = !v.starred;
-        }
-        return v;
-      });
-      setList(updated);
-    }
-  };
-
-  const starMany = async () => {
-    const toStar = list.filter((v) => v.checked).map((i) => i.id);
-    await starItemsById((data as any).csrfmiddlewaretoken, toStar);
-    const updated = list.map((v) => {
-      if (v.checked) {
-        v.starred = true;
-      }
-      return v;
-    });
-    setList(updated);
-  };
-
-  const unstarMany = async () => {
-    const toUnstar = list.filter((v) => v.checked).map((i) => i.id);
-    await unstarItemsById((data as any).csrfmiddlewaretoken, toUnstar);
-    const updated = list.map((v) => {
-      if (v.checked) {
-        v.starred = false;
-      }
-      return v;
-    });
-    setList(updated);
-  };
-
-  const calculateBulkEditOptions = (items: any[]) => {
-    editOptions.starEnabled = false;
-    editOptions.unstarEnabled = false;
-    editOptions.deleteEnabled = false;
-    editOptions.unwatchEnabled = false;
-
-    items.forEach((val) => {
-      if (val.checked) {
-        !val.starred && (editOptions.starEnabled = true);
-        val.starred && (editOptions.unstarEnabled = true);
-        editOptions.deleteEnabled = true;
-        editOptions.unwatchEnabled = true;
-      }
-    });
-    setEditOptions({ ...editOptions });
-  };
-
-  const toggleItemChecked = (item) => {
-    const newList = list.map((v) => {
-      if (v.id === item.id) {
-        v.checked = !v.checked;
-      }
-      return v;
-    });
-    calculateBulkEditOptions(newList);
-    setList(newList);
-  };
-
-  const unwatchItem = async (toUnWatch) => {
-    await unwatchItemsByUrls(data.csrfmiddlewaretoken, [toUnWatch]);
-    const updated = list.filter((v) => v.id !== toUnWatch.id);
-    setList(updated);
-  };
-  const unwatchMany = async () => {
-    const toUnWatch = list.filter((v) => v.checked);
-    await unwatchItemsByUrls(data.csrfmiddlewaretoken, toUnWatch);
-    const updated = list.filter((v) => !v.checked);
-    setList(updated);
-  };
-
-  const deleteCollectionItem = async (item) => {
-    await updateDeleteCollectionItem(item, data.csrfmiddlewaretoken, true);
-    const previous = [...list];
-    const listWithDelete = list.filter((v) => v.id !== item.id);
-    setList(listWithDelete);
-    setToastData({
-      mainText: `${item.title} removed from your collection`,
-      shortText: "Article removed",
-      buttonText: "Undo",
-      buttonHandler: async () => {
-        await updateDeleteCollectionItem(item, data.csrfmiddlewaretoken, false);
-        setToastData(null);
-        setList(previous);
-      },
-    });
-  };
-
-  const deleteFrequentlyViewed = async (bookmarkData: BookmarkData) => {
-    const original = list;
-    console.log(list.map((v, idx) => `${idx} ${v.url}`));
-    const filteredEntries = list.filter(
-      (entry) => entry.url !== bookmarkData.url
-    );
-    console.log(filteredEntries.map((v, idx) => `${idx} ${v.url}`));
-    setFrequentlyUsed(filteredEntries);
-    setList(filteredEntries);
-    setToastData({
-      mainText: `${bookmarkData.title} removed`,
-      shortText: "Article removed",
-      buttonText: "UNDO",
-      buttonHandler: async () => {
-        const restored: any = {
-          url: bookmarkData.url,
-          title: bookmarkData.title,
-          timestamp: new Date(bookmarkData.created).getTime(),
-          parents: bookmarkData.parents,
-          visitCount: bookmarkData.visitCount || 1,
-        };
-        setFrequentlyUsed(original);
-        setList(original);
-        setToastData(null);
-      },
-    });
-  };
-
-  const collectionsSaveHandler = async (
-    e: React.FormEvent<HTMLFormElement> | React.BaseSyntheticEvent,
-    item: BookmarkData
-  ) => {
-    let formData;
-    const form = e.target as HTMLFormElement;
-    formData = new FormData(form);
-    console.log(item);
-    await updateCollectionItem(
-      item,
-      new URLSearchParams([...(formData as any)]),
-      data.csrfmiddlewaretoken
-    );
-  };
-
-  let cardList = list.map((item) => {
-    if (
-      currentTab === TabVariant.NOTIFICATIONS ||
-      currentTab === TabVariant.STARRED
-    ) {
-      return (
-        <NotificationCardListItem
-          handleDelete={deleteItem}
-          item={item}
-          toggleSelected={toggleItemChecked}
-          toggleStarred={toggleStarItem}
-          key={item.id}
-        />
-      );
-    } else if (currentTab == TabVariant.WATCHING) {
-      return (
-        <WatchedCardListItem
-          onUnwatched={unwatchItem}
-          item={item}
-          toggleSelected={toggleItemChecked}
-          key={item.id}
-        />
-      );
-    } else if (currentTab == TabVariant.COLLECTIONS) {
-      return (
-        <CollectionListItem
-          item={item}
-          onEditSubmit={collectionsSaveHandler}
-          key={item.id}
-          showEditButton={true}
-          handleDelete={deleteCollectionItem}
-        />
-      );
-    } else if (currentTab == TabVariant.FREQUENTLY_VIEWED) {
-      return (
-        <CollectionListItem
-          item={item}
-          onEditSubmit={collectionsSaveHandler}
-          key={item.id}
-          showEditButton={false}
-          handleDelete={deleteFrequentlyViewed}
-        />
-      );
-    }
-  });
-
-  return (
-    <>
-      <SearchFilter
-        filters={
-          currentTab === TabVariant.STARRED ||
-          currentTab === TabVariant.NOTIFICATIONS
-            ? FILTERS
-            : []
-        }
-        sorts={currentTab === TabVariant.WATCHING ? [] : SORTS}
-      />
-      {showSelectToolbar && (
-        <SelectedNotificationsBar
-          isChecked={selectAllChecked}
-          onStarSelected={starMany}
-          onSelectAll={(e) => {
-            const newList = list.map((item) => {
-              return { ...item, checked: e.target.checked };
-            });
-            setList(newList);
-            setSelectAllChecked(!selectAllChecked);
-            calculateBulkEditOptions(newList);
-          }}
-          onUnstarSelected={unstarMany}
-          onDeleteSelected={deleteMany}
-          buttonStates={editOptions}
-          onUnwatchSelected={unwatchMany}
-          watchedTab={currentTab === TabVariant.WATCHING}
-        />
-      )}
-      <ul className="notification-list">
-        {currentTab === TabVariant.WATCHING && (
-          <div className="icon-card-list">{cardList}</div>
-        )}
-        {currentTab !== TabVariant.WATCHING && cardList}
-      </ul>
-      {hasMore && (
-        <div className="pagination">
-          <Button
-            type="primary"
-            onClickHandler={() => {
-              setSelectAllChecked(false);
-              setOffset(offset + list.length);
-            }}
-          >
-            Show more
-          </Button>
-        </div>
-      )}
-    </>
-  );
-}
