+++ client/src/plus/notifications/watched-items-tab.tsx
@@ -1,0 +1,122 @@
+import { useState, useRef, useEffect } from "react";
+import { useWatchedItemsApiEndpoint, unwatchItemsByUrls } from "../common/api";
+import { showMoreButton } from "../common/plus-tabs";
+import WatchedCardListItem from "../icon-card";
+import SearchFilter from "../search-filter";
+import SelectedNotificationsBar from "./notification-select";
+
+export function WatchedTab({ selectedTerms, selectedFilter, selectedSort }) {
+  const [offset, setOffset] = useState(0);
+  const [selectAllChecked, setSelectAllChecked] = useState(false);
+  const [list, setList] = useState<Array<any>>([]);
+  const listRef = useRef<Array<any>>([]);
+
+  const { data, error, isLoading, hasMore } = useWatchedItemsApiEndpoint(
+    offset,
+    selectedTerms,
+    selectedFilter,
+    selectedSort
+  );
+
+  listRef.current = list;
+
+  useEffect(() => {
+    if (data && !!data.items) {
+      setList([
+        ...listRef.current,
+        ...data.items.map((item) => {
+          return { ...item, checked: false };
+        }),
+      ]);
+    }
+  }, [data]);
+
+  useEffect(() => {
+    setSelectAllChecked(false);
+    setList([]);
+    setOffset(0);
+  }, [selectedFilter, selectedSort, selectedTerms]);
+
+  const [editOptions, setEditOptions] = useState({
+    starEnabled: false,
+    unstarEnabled: false,
+    deleteEnabled: false,
+    unwatchEnabled: false,
+  });
+
+  const calculateBulkEditOptions = (items: any[]) => {
+    editOptions.starEnabled = false;
+    editOptions.unstarEnabled = false;
+    editOptions.deleteEnabled = false;
+    editOptions.unwatchEnabled = false;
+
+    items.forEach((val) => {
+      if (val.checked) {
+        !val.starred && (editOptions.starEnabled = true);
+        val.starred && (editOptions.unstarEnabled = true);
+        editOptions.deleteEnabled = true;
+        editOptions.unwatchEnabled = true;
+      }
+    });
+    setEditOptions({ ...editOptions });
+  };
+
+  const toggleItemChecked = (item) => {
+    const newList = list.map((v) => {
+      if (v.id === item.id) {
+        v.checked = !v.checked;
+      }
+      return v;
+    });
+    calculateBulkEditOptions(newList);
+    setList(newList);
+  };
+  const unwatchMany = async () => {
+    const toUnWatch = list.filter((v) => v.checked);
+    await unwatchItemsByUrls(data.csrfmiddlewaretoken, toUnWatch);
+    const updated = list.filter((v) => !v.checked);
+    setList(updated);
+  };
+
+  const unwatchItem = async (toUnWatch) => {
+    await unwatchItemsByUrls(data.csrfmiddlewaretoken, [toUnWatch]);
+    const updated = list.filter((v) => v.id !== toUnWatch.id);
+    setList(updated);
+  };
+
+  return (
+    <>
+      <SearchFilter filters={[]} sorts={[]} />
+      <SelectedNotificationsBar
+        isChecked={selectAllChecked}
+        onStarSelected={null}
+        onSelectAll={(e) => {
+          const newList = list.map((item) => {
+            return { ...item, checked: e.target.checked };
+          });
+          setList(newList);
+          setSelectAllChecked(!selectAllChecked);
+          calculateBulkEditOptions(newList);
+        }}
+        onUnstarSelected={null}
+        onDeleteSelected={null}
+        buttonStates={editOptions}
+        onUnwatchSelected={unwatchMany}
+        watchedTab={true}
+      />
+      <ul className="notification-list">
+        <div className="icon-card-list">
+          {list.map((item) => (
+            <WatchedCardListItem
+              onUnwatched={unwatchItem}
+              item={item}
+              toggleSelected={toggleItemChecked}
+              key={item.id}
+            />
+          ))}
+        </div>
+      </ul>
+      {hasMore && showMoreButton(setSelectAllChecked, setOffset, list)}
+    </>
+  );
+}
